Use actions not context menus
