+++ app/context-menus.js
@@ -9,17 +9,6 @@ const {
 const path = require('path')
 const pathPosix = path.posix
 
-// For desktop shortcuts
-const createDesktopShortcut = require('create-desktop-shortcuts')
-
-const faviconFinder = require('favicon')
-const webPConverter = require('webp-converter')
-const svgToPng = require('svg-to-png')
-const pngToIco = require('png-to-ico')
-const fs = require('fs')
-const sharp = require('sharp')
-// Kyran: I don't like the number of packages we're using just to make shortcuts
-
 module.exports = {
   attachContextMenus
 }
@@ -58,7 +47,6 @@ function attachContextMenus ({ window, createWindow }) {
       linkGroup(params),
       saveGroup(params),
       editGroup(params),
-      pageGroup(window.web || window.webContents),
       developmentGroup(window.web || window.webContents, params)
     ])
   }
@@ -152,80 +140,6 @@ function attachContextMenus ({ window, createWindow }) {
       new MenuItem({
         label: 'Hard Reload',
         click: wc.reloadIgnoringCache
-      })
-    ]
-  }
-
-  function pageGroup (wc) {
-    return [
-      new MenuItem({
-        label: 'Create shortcut',
-        click: async () => {
-          const outputPath = (await dialog.showOpenDialog({
-            properties: ['openDirectory']
-          })).filePaths[0]
-          const appPath = process.argv[0]
-
-          const shortcutName = wc.getTitle().replace(/[\/|\\:*?"<>| ]/g, '') // Kyran: Normalise into possible file name, maybe we can do this nicer. We get rid of spaces because FS issues.
-
-          const shortcut = {
-            filePath: appPath,
-            outputPath: outputPath,
-            name: shortcutName,
-            comment: 'Agregore Browser',
-            arguments: wc.getURL()
-          }
-
-          createShortcut = icon => {
-            if(icon) shortcut.icon = icon
-            // TODO: Kyran: Use Agregore icon if no icon provided.
-            // TODO: Kyran: OSX doesn't have arguments option. See https://github.com/RangerMauve/agregore-browser/pull/53#issuecomment-705654060 for solution.
-            createDesktopShortcut({
-              windows: shortcut,
-              linux: shortcut
-            })
-          }
-
-          let faviconURL
-          try {
-            faviconURL = await wc.executeJavaScript('document.querySelector("link[rel*=\'icon\']").href')
-          } catch {}
-          try {
-            if (!faviconURL) throw 'No favicon'
-            wc.session.on('will-download', (event, item, webContents) => {
-              if (item.getURL() === faviconURL) {
-                const savePath = path.join(app.getPath('userData'), 'PWAs', shortcutName, '/')
-                const savePathNamed = savePath + 'favicon'
-                const savePathDownload = savePathNamed + item.getFilename().replace(/.*(\.[a-z]*)/i, '$1')
-
-                item.setSavePath(savePathDownload)
-                item.once('done', async () => {
-                  try {
-                    try {
-                      await sharp(savePathDownload).png().resize(256, 256).toFile(savePathNamed + '.png')
-                    } catch {
-                      await webPConverter.cwebp(savePathDownload, savePathNamed + '.webp') // TODO: Kyran: Delete when done
-                      await webPConverter.dwebp(savePathNamed + '.webp', savePathNamed + '.png', '-o') // TODO: Kyran: Delete when done if Windows
-                    }
-                    const iconType = (process.platform === 'win32') ? '.ico' : '.png'
-                    if (iconType === '.ico') {
-                      const buffer = await pngToIco(savePathNamed + '.png')
-                      fs.writeFileSync(savePathNamed + '.ico', buffer)
-                    }
-                    createShortcut(savePathNamed + iconType)
-                  } catch (error) {
-                    console.log(error)
-                    createShortcut()
-                  }
-                })
-              }
-            })
-            wc.downloadURL(faviconURL)
-          } catch (error) {
-            console.log(error)
-            createShortcut()
-          }
-        }
       })
     ]
   }
